# docs/openapi.yaml
openapi: 3.0.1
info:
  title: MOS (Meal and Reminder System) API
  description: API for managing meal requirements, reminders, and user authentication
  version: 1.0.0
  contact:
    name: MOS Development Team
    email: dev@mos.example.com
servers:
- url: http://localhost:8080
  description: Development server
- url: https://api.mos.example.com
  description: Production server
paths:
...
  /api/user-management/caretakers:
    post:
      tags:
      - user-management
      summary: Create a caretaker
      description: Create a new caretaker user and assign them to the current caregiver
      operationId: createCaretaker
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaretakerRequest"
            example:
              name: "John Doe"
              email: "john.doe@example.com"
      responses:
        "200":
          description: Caretaker created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "550e8400-e29b-41d4-a716-446655440002"
                email: "john.doe@example.com"
                displayName: "John Doe"
                userType: "RESIDENT"
                isActive: true
                createdAt: "2025-10-07T11:35:30.000Z"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
                  message:
                    type: string
                    example: "Email already exists"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"

    get:
      tags:
      - user-management
      summary: List caretakers
      description: Get all caretakers assigned to the current caregiver
      operationId: getCaretakers
      security:
      - bearerAuth: []
      responses:
        "200":
          description: List of caretakers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
              example:
                - id: "550e8400-e29b-41d4-a716-446655440002"
                  email: "john.doe@example.com"
                  displayName: "John Doe"
                  userType: "RESIDENT"
                  isActive: true
                  createdAt: "2025-10-07T11:35:30.000Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
  /api/user-management/caretakers/{caretakerId}/permissions:
    get:
      tags:
      - user-management
      summary: Get caretaker permissions
      description: Get all permissions for a specific caretaker
      operationId: getCaretakerPermissions
      security:
      - bearerAuth: []
      parameters:
      - name: caretakerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Caretaker ID
      responses:
        "200":
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["CREATE_REMINDERS", "VIEW_REMINDERS", "MEAL_REQUIREMENTS"]
        "403":
          description: Forbidden - Caretaker not assigned to current caregiver
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "Access denied"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"

    put:
      tags:
      - user-management
      summary: Set caretaker permissions
      description: Set permissions for a specific caretaker
      operationId: setCaretakerPermissions
      security:
      - bearerAuth: []
      parameters:
      - name: caretakerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Caretaker ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPermissionsRequest"
            example:
              permissions: ["CREATE_REMINDERS", "VIEW_REMINDERS"]
      responses:
        "200":
          description: Permissions updated successfully
        "403":
          description: Forbidden - Caretaker not assigned to current caregiver
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "Access denied"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
  /api/user-management/caretakers/{caretakerId}:
    delete:
      tags:
      - user-management
      summary: Delete caretaker
      description: Delete a caretaker and remove all their assignments
      operationId: deleteCaretaker
      security:
      - bearerAuth: []
      parameters:
      - name: caretakerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Caretaker ID
      responses:
        "200":
          description: Caretaker deleted successfully
        "403":
          description: Forbidden - Caretaker not assigned to current caregiver
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "Access denied"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
  /api/user-management/permissions:
    get:
      tags:
      - user-management
      summary: Get current user permissions
      description: Get all permissions for the current authenticated user
      operationId: getCurrentUserPermissions
      security:
      - bearerAuth: []
      responses:
        "200":
          description: List of current user permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["CREATE_REMINDERS", "VIEW_REMINDERS", "MEAL_REQUIREMENTS"]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        displayName:
          type: string
          description: User display name
          example: "John Doe"
        userType:
          type: string
          enum: [ADMIN, CAREGIVER, RESIDENT]
          description: User role
          example: "RESIDENT"
        isActive:
          type: boolean
          description: Whether user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-10-07T11:35:30.000Z"
    CreateCaretakerRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Caretaker's display name
          example: "John Doe"
        email:
          type: string
          format: email
          description: Caretaker's email address
          example: "john.doe@example.com"
    SetPermissionsRequest:
      type: object
      required:
        - permissions
      properties:
        permissions:
          type: array
          items:
            type: string
            enum: [CREATE_REMINDERS, VIEW_REMINDERS, MEAL_REQUIREMENTS, MEAL_SUGGESTIONS, STATISTICS]
          description: List of permissions to grant
          example: ["CREATE_REMINDERS", "VIEW_REMINDERS", "MEAL_REQUIREMENTS"]